requirements.txt
flask==3.0.0
twilio==8.10.0
anthropic>=0.40.0
apscheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
Werkzeug==3.0.1
Pillow>=10.0.0              # image resizing before sending to Ollama (prevents llama runner OOM crash)
pdfplumber>=0.10.0         # PDF text extraction for statement parsing (falls back to AI if unavailable)
prometheus-flask-exporter  # exposes /metrics endpoint for Prometheus scraping
